requirements.txt
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
rank-bm25>=0.2.2
transformers>=4.30.0
wikipedia-api>=0.8.0
beautifulsoup4>=4.12.0
nltk>=3.8.0
rouge-score>=0.1.2
bert-score>=0.3.10
scikit-learn>=1.3.0
streamlit>=1.20.0
pandas>=2.0.0
numpy>=1.24.0
torch>=2.0.0
requests>=2.31.0
tqdm>=4.66.0
pydantic>=2.0.0
pypdf>=3.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
